make use of _is_zero() consistent
