+++ odetoolbox/system_of_shapes.py
@@ -27,20 +27,10 @@ import sympy
 import sympy.matrices
 
 from .shapes import Shape
+from .sympy_printer import _is_zero
 
 
-def _is_zero(x):
-    r"""
-    In the ideal case, we would like to use sympy.simplify() to do simplification of an expression before comparing it to zero. However, for expressions of moderate size (e.g. a few dozen terms involving exp() functions), it becomes unbearably slow. We therefore use this internal function, so that the simplification function can be easily switched over.
-
-    Tests by expand_mul only, suitable for polynomials and rational functions.
-
-    Ref.: https://github.com/sympy/sympy PR #13877 by @normalhuman et al. merged on Jan 27, 2018
-    """
-    return bool(sympy.expand_mul(x).is_zero)
-
-
-class SystemOfShapes(object):
+class SystemOfShapes:
     r"""
     Represent a dynamical system in the canonical form :math:`\mathbf{x}' = \mathbf{Ax} + \mathbf{c}`.
     """
